+++ src/VBox/Main/src-server/linux/NetIf-linux.cpp
@@ -4,7 +4,7 @@
  */
 
 /*
- * Copyright (C) 2008-2024 Oracle and/or its affiliates.
+ * Copyright (C) 2008-2025 Oracle and/or its affiliates.
  *
  * This file is part of VirtualBox base platform packages, as
  * available from https://www.virtualbox.org.
@@ -254,7 +254,7 @@ static int getInterfaceInfo(int iSocket, const char *pszName, PNETIFINFO pInfo)
                     pInfo->IPv6Address.au32[1] = htonl(IPv6Address.au32[1]);
                     pInfo->IPv6Address.au32[2] = htonl(IPv6Address.au32[2]);
                     pInfo->IPv6Address.au32[3] = htonl(IPv6Address.au32[3]);
-                    RTNetPrefixToMaskIPv6(uLength, &pInfo->IPv6NetMask);
+                    RTNetPrefixToMaskIPv6((int)uLength, &pInfo->IPv6NetMask);
                 }
             }
             fclose(fp);
